sensors_api/requirements.txt
fastapi~=0.115.0
python-decouple~=3.8
sqlmodel~=0.0.22
uvicorn~=0.31.1